Do not create symlink on install
When using the "install" make target, do not symlink libnss_altfiles.so.2
to libnss_altfiles.so: glibc2 and eglibc will directly load the .2-suffixed
shared object without needing the symlink. The .so.1/.so suffixes would be
for supporting glibc1, but this NSS module is not expected to work with such
and older libc. Therefore, the symlink can just be dropped.