Fix linking building for 32-bit in a 64-bit host
Links the shared library using -nostdlib, which is safe because the
undefined symbols would be available later on when the shared library
is loaded as a plugin by the libc. Yet, -lpthread is kept to ensure
that the mutex will work as expected.

Also, -as-needed is passed to the linker to avoid including unneeded
dependencies in the ELF header of the shared library.