Support setting the NSS file types to support
Add a --with-types=FOO flag to "configure", with FOO being a comma-separated
list of file types to support, and use its value to determine which files to
build (in Makefile).

To maintain the original behavior of the module, the default value is set
to "pwd,grp".

Using "--with-types=all" will select all file types.